# backend/docs/api/phase-one-frontend.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "山海巡厄录第一阶段前台 API 契约",
    "version": "1.0.0",
    "description": "基于 backend/routes/api.php、ApiRequest、Resource 与现有 Feature Tests 收口的第一阶段前台 API 正式契约。所有业务失败当前仍返回 HTTP 200，并通过 code/message/data=null 表达业务错误。"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/api/characters": {
      "get": {
        "tags": [
          "Character"
        ],
        "summary": "获取当前用户角色列表",
        "operationId": "listCharacters",
        "responses": {
          "200": {
            "description": "成功返回当前用户全部角色列表；失败仍使用统一错误外壳。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CharacterListSuccessResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 0,
                      "message": "ok",
                      "data": {
                        "characters": [
                          {
                            "character_id": 1001,
                            "character_name": "青山",
                            "class_id": "class_jingang",
                            "class_name": "金刚",
                            "level": 1,
                            "exp": 0,
                            "is_active": 1,
                            "unspent_stat_points": 0,
                            "added_strength": 0,
                            "added_mana": 0,
                            "added_constitution": 0,
                            "added_dexterity": 0,
                            "long_term_growth_stage": null
                          },
                          {
                            "character_id": 1002,
                            "character_name": "白露",
                            "class_id": "class_fashi",
                            "class_name": "法师",
                            "level": 1,
                            "exp": 0,
                            "is_active": 0,
                            "unspent_stat_points": 0,
                            "added_strength": 0,
                            "added_mana": 0,
                            "added_constitution": 0,
                            "added_dexterity": 0,
                            "long_term_growth_stage": null
                          }
                        ]
                      }
                    }
                  },
                  "unauthorized": {
                    "value": {
                      "code": 10002,
                      "message": "未登录或登录失效",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Character"
        ],
        "summary": "创建角色",
        "operationId": "createCharacter",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "class_id",
                  "character_name"
                ],
                "properties": {
                  "class_id": {
                    "type": "string",
                    "description": "职业 ID"
                  },
                  "character_name": {
                    "type": "string",
                    "description": "角色名称"
                  }
                }
              },
              "example": {
                "class_id": "class_jingang",
                "character_name": "联调烟测角"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功返回角色与初始化后的固定 12 槽位；失败仍使用统一错误外壳。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CharacterCreateSuccessResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 0,
                      "message": "ok",
                      "data": {
                        "character": {
                          "character_id": 1002,
                          "character_name": "联调烟测角",
                          "class_id": "class_jingang",
                          "class_name": "金刚",
                          "level": 1,
                          "exp": 0,
                          "is_active": 0,
                          "unspent_stat_points": 0,
                          "added_strength": 0,
                          "added_mana": 0,
                          "added_constitution": 0,
                          "added_dexterity": 0,
                          "long_term_growth_stage": null
                        },
                        "equipment_slots": [
                          {
                            "slot_key": "main_weapon",
                            "equipped_instance_id": null,
                            "equipment": null
                          },
                          {
                            "slot_key": "sub_weapon",
                            "equipped_instance_id": null,
                            "equipment": null
                          },
                          {
                            "slot_key": "armor",
                            "equipped_instance_id": null,
                            "equipment": null
                          }
                        ]
                      }
                    }
                  },
                  "invalid_params": {
                    "value": {
                      "code": 10001,
                      "message": "请求参数不合法",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/characters/{character_id}/activate": {
      "post": {
        "tags": [
          "Character"
        ],
        "summary": "切换当前启用角色",
        "operationId": "activateCharacter",
        "parameters": [
          {
            "$ref": "#/components/parameters/character_id"
          }
        ],
        "responses": {
          "200": {
            "description": "成功后返回当前启用角色摘要；失败仍使用统一错误外壳。",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CharacterDetailSuccessResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 0,
                      "message": "ok",
                      "data": {
                        "character": {
                          "character_id": 1002,
                          "character_name": "白露",
                          "class_id": "class_fashi",
                          "class_name": "法师",
                          "level": 1,
                          "exp": 0,
                          "is_active": 1,
                          "unspent_stat_points": 0,
                          "added_strength": 0,
                          "added_mana": 0,
                          "added_constitution": 0,
                          "added_dexterity": 0,
                          "long_term_growth_stage": null
                        }
                      }
                    }
                  },
                  "forbidden": {
                    "value": {
                      "code": 10102,
                      "message": "无权访问该角色",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用当前用户 api_token 作为 Bearer Token。"
      }
    },
    "parameters": {
      "character_id": {
        "name": "character_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
...
    },
    "schemas": {
      "ApiSuccessEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "const": 0
          },
          "message": {
            "type": "string",
            "const": "ok"
          },
          "data": {}
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "null"
          }
        }
      },
      "CharacterSummary": {
        "type": "object",
        "required": [
          "character_id",
          "character_name",
          "class_id",
          "class_name",
          "level",
          "exp",
          "is_active",
          "unspent_stat_points",
          "added_strength",
          "added_mana",
          "added_constitution",
          "added_dexterity",
          "long_term_growth_stage"
        ],
        "properties": {
          "character_id": {
            "type": "integer"
          },
          "character_name": {
            "type": "string"
          },
          "class_id": {
            "type": "string"
          },
          "class_name": {
            "type": "string"
          },
          "level": {
            "type": "integer"
          },
          "exp": {
            "type": "integer"
          },
          "is_active": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "unspent_stat_points": {
            "type": "integer"
          },
          "added_strength": {
            "type": "integer"
          },
          "added_mana": {
            "type": "integer"
          },
          "added_constitution": {
            "type": "integer"
          },
          "added_dexterity": {
            "type": "integer"
          },
          "long_term_growth_stage": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "EquipmentInstanceSummary": {
        "type": "object",
        "required": [
          "equipment_instance_id",
          "item_id",
          "item_name",
          "equipment_slot",
          "rarity",
          "icon",
          "bind_type",
          "enhance_level",
          "durability",
          "max_durability"
        ],
        "properties": {
          "equipment_instance_id": {
            "type": "integer"
          },
          "item_id": {
            "type": "string"
          },
          "item_name": {
            "type": "string"
          },
          "equipment_slot": {
            "type": "string"
          },
          "rarity": {
            "type": "string"
          },
          "icon": {
            "type": [
              "string",
              "null"
            ]
          },
          "bind_type": {
            "type": "string"
          },
          "enhance_level": {
            "type": "integer"
          },
          "durability": {
            "type": "integer"
          },
          "max_durability": {
            "type": "integer"
          }
        }
      },
      "EquipmentSlot": {
        "type": "object",
        "required": [
          "slot_key",
          "equipped_instance_id",
          "equipment"
        ],
        "properties": {
          "slot_key": {
            "type": "string"
          },
          "equipped_instance_id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "equipment": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/EquipmentInstanceSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "CharacterCreateData": {
        "type": "object",
        "required": [
          "character",
          "equipment_slots"
        ],
        "properties": {
          "character": {
            "$ref": "#/components/schemas/CharacterSummary"
          },
          "equipment_slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EquipmentSlot"
            }
          }
        }
      },
      "CharacterDetailData": {
        "type": "object",
        "required": [
          "character"
        ],
        "properties": {
          "character": {
            "$ref": "#/components/schemas/CharacterSummary"
          }
        }
      },
      "CharacterListData": {
        "type": "object",
        "required": [
          "characters"
        ],
        "properties": {
          "characters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CharacterSummary"
            }
          }
        }
      },
...
      "CharacterCreateSuccessResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiSuccessEnvelope"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/CharacterCreateData"
              }
            }
          }
        ]
      },
      "CharacterDetailSuccessResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiSuccessEnvelope"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/CharacterDetailData"
              }
            }
          }
        ]
      },
      "CharacterListSuccessResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiSuccessEnvelope"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/CharacterListData"
              }
            }
          }
        ]
      },
...
    }
  }
}